// client/dist/widgets/common/print/src/config.ts
export enum PrintResultState {
  Loading = 'LOADING',
  Success = 'SUCCESS',
  Error = 'ERROR'
}

export enum Views {
  PrintTemplate = 'PRINT TEMPLATE',
  PrintResult = 'PRINT RESULT'
}

export interface PrintResultListItemTyle {
  resultId: string
  url: string
  title: string
  state: PrintResultState
}

export interface JimuMapViews { [viewId: string]: JimuMapView }
export type IMJimuMapViews = ImmutableObject<JimuMapViews>

export type IMPrintResultListItemTyle = ImmutableObject<PrintResultListItemTyle>